Initial commit for pts-automation-tool
This initial commit sets up the foundation for the pts-automation-tool
project. It includes essential files such as README.md, which provides
an overview and instructions for the project, and Kconfig, which offers
a user-friendly configuration menu for test parameters.

The Makefile is included to automate build processes with targets for
configuring, saving, and displaying test results. Additionally, the
commit adds a config file for default settings, comprehensive documentation
for detailed guidance, and scripts like pts_nginx.sh and pts_pgbench.sh
for benchmarking with Nginx and pgbench, respectively.

The saveresult script is also included to ensure test results are saved
systematically. This setup aims to streamline performance testing using
the Phoronix Test Suite, providing a robust starting point for further
development.

Signed-off-by: Hemanth Selam <Hemanth.Selam@amd.com>