Bug fix in config file and adding expect package support
This commit addresses a critical bug in the configuration script
and introduces support for the expect package to enhance automation
capabilities.

The bug fix involves correcting the sed command to ensure that
previous pgbench options are properly cleared from the .config file,
allowing the pgbench submenu to display and function correctly.

Additionally, the integration of the expect package provides robust
scripting capabilities, particularly useful for handling interactive
commands. These improvements significantly enhance the functionality
and reliability of the pts-automation-tool, ensuring a smoother user
experience and more accurate configuration management.

Signed-off-by: Hemanth Selam <Hemanth.Selam@amd.com>