Added support of stress-ng test and bugfix for showresult
This patch introduces the stress-ng test and fixes a bug in the
showresult script. Additionally, it renames the test files by
removing the .sh extension for consistency and simplicity.

The Kconfig and Makefile have been updated to reflect these changes,
and the config file has been adjusted to enhance compatibility and
performance. The documentation has also been revised to provide clearer
instructions and updated information. These updates collectively aim to
improve the functionality and maintainability of the testing framework.

Signed-off-by: Hemanth Selam <Hemanth.Selam@amd.com>