Add support of sysbench test
This commit introduces support for the pts/sysbench test and
includes several updates to the testing framework.

The Kconfig file has been modified to include configuration
options for enabling and configuring the sysbench test.

The Makefile has been updated to incorporate targets for building
and running the sysbench test. The config file has been adjusted to
support sysbench benchmarking, and the documentation has been revised
to provide instructions for running the sysbench test.

Signed-off-by: Hemanth Selam <Hemanth.Selam@amd.com>