Makefile: Add support of check functionality of PTS install
Added a check to verify if the Phoronix Test Suite is already
installed, ensuring smoother setup processes. Additionally,
fixed a bug in the Makefile related to the installation parts for
openAnolis and openEuler distributions.

The support for the epel-release package has been removed for both
openAnolis and openEuler to streamline the installation process.
These updates enhance the reliability and compatibility of the testing
framework across different environments

Signed-off-by: Hemanth Selam <Hemanth.Selam@amd.com>